+++ src/string.ts
@@ -22,7 +22,7 @@ export const suffix =
   <S extends string>(s: S): `${S}${Suffix}` =>
     `${s}${suffix}`
 
-export const intercalate =
+export const infix =
   <M extends string>(middle: M) =>
   <S0 extends string, S1 extends string>(s0: S0, s1: S1): `${S0}${M}${S1}` =>
     `${s0}${middle}${s1}`